requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
email-validate==1.1.2
filelock==3.13.3
Flask==3.0.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
psycopg2-binary==2.9.6
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
SQLAlchemy==2.0.29
typing_extensions==4.10.0
validate-email==1.3
Werkzeug==3.0.1